python routes refactoring
